ac3d editor can read and write floating point material shi field even though it is an int in the standard so make accc read float but output int
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@7138 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: 7785d217b31cfd68250e6cb551eaf5bdff14c538
Former-commit-id: bae8ac5483f249e1f58d0fc069eebaf8e4524941

+++ src/tools/accc/accc.h
@@ -218,7 +218,7 @@ typedef struct mat
     color_t amb;
     color_t emis;
     color_t spec;
-    int shi;
+    double shi;
     double trans;
     struct mat * next;
 } mat_t;
